+++ packages/components/src/Cards/SidebarCard/index.tsx
@@ -3,9 +3,10 @@ import styled from "@emotion/styled";
 import { css, Global } from "@emotion/react";
 import { Icon } from "@iconify/react";
 import { useHass, useHash } from "@hakit/core";
-import { TimeCard, WeatherCard, Row, Column } from "@components";
+import { TimeCard, WeatherCard, Row, Column, fallback } from "@components";
 import { motion, AnimatePresence } from "framer-motion";
 import type { WeatherCardProps, TimeCardProps } from "@components";
+import { ErrorBoundary } from "react-error-boundary";
 
 const StyledTimeCard = styled(TimeCard)<{
   open: boolean;
@@ -209,17 +210,23 @@ export interface MenuItem {
   onClick: (event: React.MouseEvent<HTMLLIElement>) => void;
 }
 
-/** This component is a nice way of organizing components / groups into an easy to navigate sidebar, the "Room Cards" will automatically insert into the sidebar items if they're present on the page, eg if you have 6 RoomCards, all 6 items will be added to the sidebar automatically, you can also add your own menu items to the start of the list, this all needs a bit more thought but for now it is pretty useful! The TimeCard and WeatherCard are integrate and themed slightly different in the sidebar, if the sidebar is present, the RoomCard will only expand to the available space and not cover the sidebar */
 export interface SidebarCardProps {
+  /** should the time card be included by default @default true */
   includeTimeCard?: boolean;
+  /** should the sidebar start opened @default true */
   startOpen?: boolean;
+  /** the props for the weather card , if omitted, weather card is not rendered @default undefined */
   weatherCardProps?: WeatherCardProps;
+  /** Adding menu items can also add routes by default, disabled this if need be @default true */
   autoIncludeRoutes?: boolean;
+  /** the props for the timeCard @default { includeIcon: false, includeDate: false, center: true }*/
   timeCardProps?: TimeCardProps;
+  /** the menu items to add to the sidebar @default [] */
   menuItems?: MenuItem[];
+  /** the children to render in the sidebar */
   children?: React.ReactNode;
 }
-export function SidebarCard({
+function _SidebarCard({
   weatherCardProps,
   timeCardProps = {
     includeIcon: false,
@@ -388,5 +395,13 @@ export function SidebarCard({
         </Column>
       </StyledSidebarCard>
     </>
+  );
+}
+/** This component is a nice way of organizing components / groups into an easy to navigate sidebar, the "Room Cards" will automatically insert into the sidebar items if they're present on the page, eg if you have 6 RoomCards, all 6 items will be added to the sidebar automatically, you can also add your own menu items to the start of the list, this all needs a bit more thought but for now it is pretty useful! The TimeCard and WeatherCard are integrate and themed slightly different in the sidebar, if the sidebar is present, the RoomCard will only expand to the available space and not cover the sidebar */
+export function SidebarCard(props: SidebarCardProps) {
+  return (
+    <ErrorBoundary {...fallback({ prefix: "SidebarCard" })}>
+      <_SidebarCard {...props} />
+    </ErrorBoundary>
   );
 }