+++ packages/components/src/Shared/Modal/Modal.stories.tsx
@@ -116,9 +116,7 @@ export default {
     centered: true,
   },
 } satisfies Meta<typeof Modal>;
-export type ModalStory = StoryObj<
-  typeof ButtonCard<"light.fake_light_1", "toggle">
->;
+export type ModalStory = StoryObj<typeof ButtonCard<"light.fake_light_1">>;
 export const ModalExample: ModalStory = {
   render: Render,
   args: {