+++ packages/core/src/types/index.ts
@@ -44,6 +44,8 @@ export type HassEntityCustom = HassEntity & {
   custom: {
     /** the difference in time between the last updated value and now, eg "1 minute ago, 1 day ago etc, 5 days from now" */
     relativeTime: string;
+    /** the difference in time in milliseconds between now and the time the entity was updated / triggered */
+    timeDiff: number;
     /** if the last updated value was considered "now" */
     active: boolean;
     /** the hexColor value if the entity is a light */
@@ -66,9 +68,8 @@ type HassEntityHelper<T extends AllDomains> =
     : HassEntity;
 
 export type HassEntityWithApi<T extends AllDomains> = HassEntityCustom &
-  HassEntityHelper<T> & {
-    /** all the services associated with the domain provided, this does not require entity as the first argument */
-    api: T extends keyof SupportedServices
+  HassEntityHelper<SnakeToCamel<T>> & {
+    api: SnakeToCamel<T> extends keyof SupportedServices<"no-target">
       ? SupportedServices<"no-target">[SnakeToCamel<T>]
       : never;
   };